refactor: split types into simpler setup
Move general types into utils, and pack ENSNode types into its own namespace. Also, split the Domain and DTO types into respective namespaces.
