+++ packages/ensnode-sdk/src/indexing-status/domain/types.test.ts
@@ -1,0 +1,133 @@
+import { describe, expect, it } from "vitest";
+import { ENSNode } from "../../types";
+import { IndexingStatusDomain } from "./types";
+
+describe("Indexing Status", () => {
+  describe("Domain layer", () => {
+    it("can express a complete data model", () => {
+      // permutations defined: https://docs.google.com/spreadsheets/d/1BresRxwVBquMftKtmdRL7aayYtwy-MN0BWQMg9aufkU/edit?gid=0#gid=0
+
+      // ID: 1
+      const unhealthyRpcAndSyncQueued = {
+        chainId: 1,
+        rpcHealth: ENSNode.RPCHealth.Unhealthy,
+        indexingPhase: ENSNode.IndexingPhase.SyncQueued,
+        firstBlockToIndex: {
+          number: 123,
+        },
+        lastIndexedBlock: null,
+        lastSyncedBlock: null,
+        latestSafeBlock: null,
+      } satisfies ENSNode.RpcUnhealthyAndSyncQueued<IndexingStatusDomain.BlockInfo>;
+
+      // ID: 2
+      const healthyRpcAndSyncQueued = {
+        chainId: 2,
+        rpcHealth: ENSNode.RPCHealth.Healthy,
+        firstBlockToIndex: {
+          number: 123,
+          createdAt: new Date(),
+        },
+        indexingPhase: ENSNode.IndexingPhase.SyncQueued,
+        lastIndexedBlock: null,
+        lastSyncedBlock: null,
+        latestSafeBlock: {
+          number: 456,
+          createdAt: new Date(),
+        },
+      } satisfies ENSNode.RpcHealthyAndSyncQueued<IndexingStatusDomain.BlockInfo>;
+
+      // ID: 3
+      const unhealthyRpcAndIndexingQueued = {
+        chainId: 3,
+        rpcHealth: ENSNode.RPCHealth.Unhealthy,
+        indexingPhase: ENSNode.IndexingPhase.IndexingQueued,
+        firstBlockToIndex: {
+          number: 123,
+          createdAt: new Date(),
+        },
+        lastIndexedBlock: null,
+        lastSyncedBlock: {
+          number: 456,
+          createdAt: new Date(),
+        },
+        latestSafeBlock: null,
+      } satisfies ENSNode.RpcUnhealthyAndIndexingQueued<IndexingStatusDomain.BlockInfo>;
+
+      // ID: 4
+      const healthyRpcAndIndexingQueued = {
+        chainId: 4,
+        rpcHealth: ENSNode.RPCHealth.Healthy,
+        indexingPhase: ENSNode.IndexingPhase.IndexingQueued,
+        firstBlockToIndex: {
+          number: 123,
+          createdAt: new Date(),
+        },
+        lastIndexedBlock: null,
+        lastSyncedBlock: {
+          number: 456,
+          createdAt: new Date(),
+        },
+        latestSafeBlock: {
+          number: 789,
+          createdAt: new Date(),
+        },
+      } satisfies ENSNode.RpcHealthyAndIndexingQueued<IndexingStatusDomain.BlockInfo>;
+
+      // ID: 5
+      const unhealthyRpcAndIndexingStarted = {
+        chainId: 5,
+        rpcHealth: ENSNode.RPCHealth.Unhealthy,
+        indexingPhase: ENSNode.IndexingPhase.IndexingStarted,
+        firstBlockToIndex: {
+          number: 123,
+          createdAt: new Date(),
+        },
+        lastIndexedBlock: {
+          number: 333,
+          createdAt: new Date(),
+        },
+        lastSyncedBlock: {
+          number: 456,
+          createdAt: new Date(),
+        },
+        latestSafeBlock: null,
+      } satisfies ENSNode.RpcUnhealthyAndIndexingStarted<IndexingStatusDomain.BlockInfo>;
+
+      // ID: 6
+      const healthyRpcAndIndexingStarted = {
+        chainId: 6,
+        rpcHealth: ENSNode.RPCHealth.Healthy,
+        indexingPhase: ENSNode.IndexingPhase.IndexingStarted,
+        firstBlockToIndex: {
+          number: 123,
+          createdAt: new Date(),
+        },
+        lastIndexedBlock: {
+          number: 333,
+          createdAt: new Date(),
+        },
+        lastSyncedBlock: {
+          number: 456,
+          createdAt: new Date(),
+        },
+        latestSafeBlock: {
+          number: 789,
+          createdAt: new Date(),
+        },
+      } satisfies ENSNode.RpcHealthyAndIndexingStarted<IndexingStatusDomain.BlockInfo>;
+
+      const indexingStatus: IndexingStatusDomain.IndexingStatus = new Map();
+
+      indexingStatus.set(unhealthyRpcAndSyncQueued.chainId, unhealthyRpcAndSyncQueued);
+      indexingStatus.set(healthyRpcAndSyncQueued.chainId, healthyRpcAndSyncQueued);
+      indexingStatus.set(unhealthyRpcAndIndexingQueued.chainId, unhealthyRpcAndIndexingQueued);
+      indexingStatus.set(healthyRpcAndIndexingQueued.chainId, healthyRpcAndIndexingQueued);
+      indexingStatus.set(unhealthyRpcAndIndexingStarted.chainId, unhealthyRpcAndIndexingStarted);
+      indexingStatus.set(healthyRpcAndIndexingStarted.chainId, healthyRpcAndIndexingStarted);
+
+      // get chain by chain ID
+      expect(indexingStatus.get(healthyRpcAndSyncQueued.chainId)).toBe(healthyRpcAndSyncQueued);
+    });
+  });
+});
