+++ packages/ensnode-sdk/src/indexing-status/index.ts
@@ -1,1 +1,2 @@
-export * from "./types";
+export * from "./domain/types";
+export * from "./dto/types";
